requirements.txt
torch==2.6.0+cu126
numpy==2.1.2
streamlit==1.43.2
pillow==11.0.0